Update 11/16
Added requirement needed to meet to fight the boss. Adjusted the drop rate to be 1% to be adjusted later. When the boss requirements are met, spawns slwo enemies on the bottom half to push user to the top half, fixed bug with the instructions so that they don't overlap with one another and also so that they don't show on pause and other screens. Added a quick and simple boss hp bar as well. Looking to revamp the weapons system as currently weapon is not carried over to the boss fight.